+++ src/Jigoshop/Core/Types/ProductCategory.php
@@ -68,7 +68,7 @@ class ProductCategory implements Taxonomy
 			'query_var' => true,
 			'rewrite' => [
 				'slug' => $this->options->get('permalinks.category'),
-				'with_front' => true,
+				'with_front' => $this->options->get('permalinks.with_front'),
 				'feeds' => false,
 				'pages' => true,
 				'ep_mask' => EP_ALL,